fix: 修复numberkeyboard问题

+++ src/packages/__VUE/numberkeyboard/index.taro.vue
@@ -10,7 +10,7 @@
     <div class="nut-numberkeyboard" ref="root">
       <div class="number-board-header" v-if="title">
         <h3 class="tit">{{ title }}</h3>
-        <span class="keyboard-close" @click="closeBoard()">{{ translate('done') }}</span>
+        <span class="keyboard-close" v-if="type == 'default'" @click="closeBoard()">{{ translate('done') }}</span>
       </div>
       <div class="number-board-body">
         <div class="number-board">
@@ -61,7 +61,7 @@
               />
             </div>
           </div>
-          <div class="key-board-wrapper" @click="closeBoard()" v-if="title == ''">
+          <div class="key-board-wrapper" @click="closeBoard()">
             <div :class="['key', 'finish', { activeFinsh: clickKeyIndex == 'finish' }]">
               {{ confirmText || translate('done') }}
             </div>
@@ -159,8 +159,11 @@ export default create({
       if (customKeys.length > 2) {
         customKeys = [customKeys[0], customKeys[1]];
       }
+      if (customKeys.length == 2 && props.title && props.type != 'rightColumn') {
+        customKeys = [customKeys[0]];
+      }
       if (customKeys.length === 1) {
-        if (props.title) {
+        if (props.title && props.type != 'rightColumn') {
           keys.push({ id: customKeys[0], type: 'custom' }, { id: 0, type: 'number' }, { id: 'delete', type: 'delete' });
         } else {
           keys.push({ id: 0, type: 'number' }, { id: customKeys[0], type: 'custom' });
@@ -171,11 +174,6 @@ export default create({
           { id: 0, type: 'number' },
           { id: customKeys[1], type: 'custom' }
         );
-        if (props.title) {
-          keys.push({ id: 'delete', type: 'delete' });
-        }
-      } else {
-        keys.push({ id: 0, type: 'number' });
       }
       return keys;
     }
